Updated main and unittests
