+++ ample/util/tests/test_ample_util.py
@@ -2,8 +2,6 @@
 
 import pickle
 import os
-import shutil
-import tempfile
 import unittest
 from ample.util import ample_util
 from ample.constants import AMPLE_PKL, SHARE_DIR
@@ -30,59 +28,6 @@ class Test(unittest.TestCase):
         d = ample_util.amoptd_fix_path(optd, newroot='/foo/bar')
         self.assertEqual(d['fasta'], '/foo/bar/ampl_.fasta')
         self.assertEqual(d['mrbump_results'][0]['PHASER_logfile'], '/foo/bar/MRBUMP/search_c1_t69_r3_polyAla_mrbump/data/loc0_ALL_c1_t69_r3_polyAla/unmod/mr/phaser/phaser_loc0_ALL_c1_t69_r3_polyAla_UNMOD.log')
-
-    def test_extract_and_validate_models(self):
-        mdir = os.path.join(self.testfiles_dir, 'models')
-        amoptd = {'models': mdir,
-                  'models_dir': None}
-        sequence = 'QPRRKLCILHRNPGRCYDKIPAFYYNQKKKQCERFDWSGCGGNSNRFKTIEECRRTCIG'
-        models = ample_util.extract_and_validate_models(amoptd=amoptd, sequence=sequence, single=True)
-        self.assertEqual(len(models), 30)
-        self.assertEqual(amoptd['models_dir'], mdir)
-
-    def test_extract_and_validate_models_quark(self):
-        temp_dir = tempfile.mkdtemp(dir=os.getcwd())
-        amoptd = {'models' : os.path.join(self.testfiles_dir, 'result.tar.bz2'),
-                  'models_dir' : temp_dir,
-                  'work_dir' : '.'}
-        sequence = 'QPRRKLCILHRNPGRCYDKIPAFYYNQKKKQCERFDWSGCGGNSNRFKTIEECRRTCIG'
-        models = ample_util.extract_and_validate_models(amoptd=amoptd, sequence=sequence, single=True)
-        self.assertTrue(amoptd['quark_models'])
-        self.assertEqual(len(models), 200)
-        shutil.rmtree(temp_dir)
-        
-    def test_extract_and_validate_models_quark_zip(self):
-        temp_dir = tempfile.mkdtemp(dir=os.getcwd())
-        amoptd = {'models' : os.path.join(self.testfiles_dir, 'result.zip'),
-                  'models_dir' : temp_dir,
-                  'work_dir' : '.'}
-        sequence = 'QPRRKLCILHRNPGRCYDKIPAFYYNQKKKQCERFDWSGCGGNSNRFKTIEECRRTCIG'
-        models = ample_util.extract_and_validate_models(amoptd=amoptd, sequence=sequence, single=True)
-        self.assertTrue(amoptd['quark_models'])
-        self.assertEqual(len(models), 200)
-        shutil.rmtree(temp_dir)
-        
-    def test_extract_and_validate_models_quark_file(self):
-        temp_dir = tempfile.mkdtemp(dir=os.getcwd())
-        amoptd = {'models' : os.path.join(self.testfiles_dir, 'alldecoy.pdb'),
-                  'models_dir' : temp_dir,
-                  'work_dir' : '.'}
-        sequence = 'QPRRKLCILHRNPGRCYDKIPAFYYNQKKKQCERFDWSGCGGNSNRFKTIEECRRTCIG'
-        models = ample_util.extract_and_validate_models(amoptd=amoptd, sequence=sequence, single=True)
-        self.assertTrue(amoptd['quark_models'])
-        self.assertEqual(len(models), 10)
-        shutil.rmtree(temp_dir)
-
-    def test_extract_and_validate_models_quark_old(self):
-        temp_dir = tempfile.mkdtemp(dir=os.getcwd())
-        amoptd = {'models' : os.path.join(self.testfiles_dir, 'decoys.tar.gz'),
-                  'models_dir' : temp_dir,
-                  'work_dir' : '.'}
-        sequence = 'QPRRKLCILHRNPGRCYDKIPAFYYNQKKKQCERFDWSGCGGNSNRFKTIEECRRTCIG'
-        models = ample_util.extract_and_validate_models(amoptd=amoptd, sequence=sequence, single=True)
-        self.assertTrue(amoptd['quark_models'])
-        self.assertEqual(len(models), 200)
-        shutil.rmtree(temp_dir)
 
     def test_extract_tar(self):
         tarfile = os.path.join(self.testfiles_dir, 'result.tar.bz2')
