+++ ample/util/tests/test_pdb_edit.py
@@ -526,19 +526,6 @@ ANISOU   26  CD1BILE A   3     4035   3461   2132   1269   -829   -356       C
                           False, False, False, True, False, False, False, False, True, True, True, False, True])
         self.assertEqual(info.numAtoms(modelIdx=0), 1208)
 
-    def test_check_pdbs_1(self):
-        pdbs = glob.glob(os.path.join(self.testfiles_dir, "models", "*.pdb"))
-        self.assertTrue(pdb_edit.check_pdbs(pdbs))
-
-    def test_check_pdbs_2(self):
-        pdbs = glob.glob(os.path.join(self.testfiles_dir, "models", "*.pdb"))
-        self.assertFalse(pdb_edit.check_pdbs(pdbs, single=True, sequence="AABBCC"))
-
-    def test_check_pdbs_3(self):
-        pdbs = glob.glob(os.path.join(self.testfiles_dir, "models", "*.pdb"))
-        pdbs += [os.path.join(self.testfiles_dir, "1GU8.pdb")]
-        self.assertFalse(pdb_edit.check_pdbs(pdbs, single=True, sequence="AABBCC"))
-
     def test_select_residues_1(self):
         pdbin = os.path.join(self.testfiles_dir, "4DZN.pdb")
         pdbout = "testSelectResidues1.pdb"
@@ -789,17 +776,7 @@ ANISOU   26  CD1BILE A   3     4035   3461   2132   1269   -829   -356       C
         self.assertEqual(info.numAtoms(modelIdx=0), 1263)
 
         return
-
-    def testCheckPdbs(self):
-        pdbs = glob.glob(os.path.join(self.testfiles_dir, "models", "*.pdb"))
-        self.assertTrue(pdb_edit.check_pdbs(pdbs))
-
-        self.assertFalse(pdb_edit.check_pdbs(pdbs, single=True, sequence="AABBCC"))
-
-        pdbs += [os.path.join(self.testfiles_dir, "1GU8.pdb")]
-        self.assertFalse(pdb_edit.check_pdbs(pdbs, single=True, sequence="AABBCC"))
-
-        return
+    
 
     def testSelectResidues(self):
         pdbin = os.path.join(self.testfiles_dir, "4DZN.pdb")
